Move fair carbon cycle to new module
