+++ fair/forward.py
@@ -3,7 +3,7 @@ from __future__ import division
 import inspect
 import numpy as np
 import warnings
-from scipy.optimize import root
+
 from .ancil import natural, cmip6_volcanic, cmip6_solar, historical_scaling
 from .constants import molwt, lifetime, radeff
 from .constants.general import M_ATMOS, ppm_gtc
@@ -11,47 +11,13 @@ from .defaults import carbon, thermal
 from .forcing import ozone_tr, ozone_st, h2o_st, contrails, aerosols, bc_snow,\
                                          landuse
 from .forcing.ghg import co2_log
-from .gas_cycle.gir import calculate_alpha, step_concentration
+#from .gas_cycle.gir import calculate_alpha, step_concentration
+from .gas_cycle.fair1 import carbon_cycle
 
 
-def iirf_interp(alp_b,a,tau,iirf_h,targ_iirf):
-    """Interpolation function for finding alpha, the CO2 decay time constant
-    scaling factor, in iirf_h equation. See Eq. (7) of Millar et al ACP (2017).
+# TODO: unified interface to the different carbon cycles
 
-    Inputs:
-        alp_b    : Guess for alpha, the scale factor, for tau 
-        a        : partition fractions for CO2 boxes
-        tau      : time constants for CO2 boxes
-        iirf_h   : time horizon for time-integrated airborne fraction
-        targ_iirf: iirf_h calculated using simple parameterisation (Eq. (8),
-                   Millar et al (2017)).
-    """
 
-    iirf_arr = alp_b*(np.sum(a*tau*(1.0 - np.exp(-iirf_h/(tau*alp_b)))))
-    return iirf_arr - targ_iirf
-
-    
-def iirf_simple(c_acc, temp, r0, rc, rt, iirf_max):
-    """Simple linear iIRF relationship. Eq. (8) of Millar et al ACP (2017).
-    
-    Inputs:
-        c_acc    : cumulative airborne carbon anomaly (GtC) since
-                   pre-industrial
-        temp     : temperature anomaly since pre-industrial
-        r0       : pre-industrial time-integrated airborne fraction (yr)
-        rc       : sensitivity of time-integrated airborne fraction to airborne
-                   carbon (yr/GtC)
-        rt       : sensitivity of time-integrated airborne fraction to
-                   temperature (yr/K)
-        iirf_max : maximum value of time-integrated airborne fraction (yr)
-    
-    Outputs:
-        iirf     : time-integrated airborne fraction of carbon (yr)
-    """
-    
-    return np.min([r0 + rc * c_acc + rt * temp, iirf_max])
-    
-    
 def calculate_q(tcrecs, d, f2x, tcr_dbl, nt):
     """If TCR and ECS are supplied, calculate the q model coefficients.
     See Eqs. (4) and (5) of Millar et al ACP (2017).
@@ -90,49 +56,6 @@ def calculate_q(tcrecs, d, f2x, tcr_dbl, nt):
     return q
     
 
-def carbon_cycle(e0, c_acc0, temp, r0, rc, rt, iirf_max, time_scale_sf0, a, tau,
-    iirf_h, carbon_boxes0, c_pi, c0, e1):
-    """Calculates CO2 concentrations from emissions.
-    
-    Inputs:
-        e0            : emissions of CO2 (GtC) in timestep t-1
-        c_acc0        : cumulative airborne carbon anomaly (GtC) since
-                        pre-industrial, timestep t-1
-        temp          : temperature anomaly above pre-industrial (K)
-        r0            : pre-industrial time-integrated airborne fraction (yr)
-        rc            : sensitivity of time-integrated airborne fraction to 
-                        airborne carbon (yr/GtC)
-        rt            : sensitivity of time-integrated airborne fraction to
-                        temperature (yr/K)
-        iirf_max      : maximum value of time-integrated airborne fraction (yr)
-        time_scale_sf0: initial guess of alpha scaling factor
-        a             : partition coefficient of carbon boxes
-        tau           : present-day decay time constants of CO2 (yr)
-        iirf_h        : time horizon for time-integrated airborne fraction (yr)
-        carbon_boxes0 : carbon stored in each atmospheric reservoir at timestep
-                        t-1 (GtC)
-        c_pi          : pre-industrial concentration of CO2, ppmv
-        c0            : concentration of CO2 in timestep t-1, ppmv
-        e1            : emissions of CO2 in timestep t, GtC
-
-    Outputs:
-        c1            : concentrations of CO2 in timestep t, ppmv
-        c_acc1        : cumulative airborne carbon anomaly (GtC) since
-                        pre-industrial, timestep t
-        carbon_boxes1 : carbon stored in each atmospheric reservoir at timestep
-                        t (GtC)
-        time_scale_sf : scale factor for CO2 decay constants
-    """
-    iirf = iirf_simple(c_acc0, temp, r0, rc, rt, iirf_max)
-    time_scale_sf = root(iirf_interp, time_scale_sf0,
-      args=(a, tau, iirf_h, iirf))['x']
-    tau_new = tau * time_scale_sf
-    carbon_boxes1 = carbon_boxes0*np.exp(-1.0/tau_new) + a*e1 / ppm_gtc
-    c1 = np.sum(carbon_boxes1) + c_pi
-    c_acc1 = c_acc0 + 0.5*(e1 + e0) - (c1 - c0)*ppm_gtc
-    return c1, c_acc1, carbon_boxes1, time_scale_sf
-    
-    
 def emis_to_conc(c0, e0, e1, ts, lt, vm):
     """Calculate concentrations of well mixed GHGs from emissions for simple
     one-box model.
@@ -278,8 +201,10 @@ def fair_scm(
     if gir_carbon_cycle:
         g1 = np.sum(a*tau * (1 - (1 + iirf_h/tau) * np.exp(-iirf_h/tau)))
         g0 = 1/(np.sinh(np.sum(a*tau*(1 - np.exp(-iirf_h/tau)) , axis=-1)/g1))
-
-### START HERE
+# TODO
+#        from .gas_cycle.gir import calculate_alpha, step_concentration
+#    else:
+#        from .gas_cycle.fair1 import calculate_alpha, step_concentration
 
     # Set up the output timeseries variables depending on options and perform
     # basic sense checks
